requirements.txt

streamlit
pdfplumber
PyMuPDF
python-docx
pandas
openpyxl
pytesseract
Pillow
